Move singletons into implementing classes soas not to pollute the static space.
