+++ controller/src/TrainingWheels/Job/ResourceJob.php
@@ -15,6 +15,9 @@ class ResourceJob extends Job {
    */
   protected function resourceCreate() {
     $course = CourseFactory::singleton()->get($this->course_id);
+    if (!$course) {
+      throw new Exception("Course with id $this->course_id does not exist.");
+    }
     $resources = empty($this->params['resources']) ? '*' : $this->params['resources'];
     $course->usersResourcesCreate($this->params['user_names'], $resources);
   }
@@ -28,6 +31,9 @@ class ResourceJob extends Job {
    */
   protected function resourceDelete() {
     $course = CourseFactory::singleton()->get($this->course_id);
+    if (!$course) {
+      throw new Exception("Course with id $this->course_id does not exist.");
+    }
     $resources = empty($this->params['resources']) ? '*' : $this->params['resources'];
     $course->usersResourcesDelete($this->params['user_names'], $resources);
   }
